move materialcache et al to material-networking library
