+++ assignment-client/CMakeLists.txt
@@ -14,7 +14,7 @@ link_hifi_libraries(
   audio avatars octree gpu graphics shaders fbx hfm entities
   networking animation recording shared script-engine embedded-webserver
   controllers physics plugins midi image
-  model-networking ktx shaders
+  material-networking model-networking ktx shaders
 )
 
 add_dependencies(${TARGET_NAME} oven)
