+++ libraries/entities-renderer/CMakeLists.txt
@@ -1,6 +1,6 @@
 set(TARGET_NAME entities-renderer)
 setup_hifi_library(Network Script)
-link_hifi_libraries(shared workload gpu shaders procedural graphics model-networking script-engine render render-utils image qml ui pointers)
+link_hifi_libraries(shared workload gpu shaders procedural graphics material-networking model-networking script-engine render render-utils image qml ui pointers)
 include_hifi_library_headers(networking)
 include_hifi_library_headers(gl)
 include_hifi_library_headers(ktx)
