+++ libraries/graphics-scripting/CMakeLists.txt
@@ -1,4 +1,4 @@
 set(TARGET_NAME graphics-scripting)
 setup_hifi_library()
-link_hifi_libraries(shared networking graphics fbx image model-networking script-engine)
+link_hifi_libraries(shared networking graphics fbx image material-networking model-networking script-engine)
 include_hifi_library_headers(gpu)
