+++ libraries/material-networking/CMakeLists.txt
@@ -1,0 +1,5 @@
+set(TARGET_NAME material-networking)
+setup_hifi_library()
+link_hifi_libraries(shared shaders networking graphics fbx ktx image gl)
+include_hifi_library_headers(gpu)
+include_hifi_library_headers(hfm)
