+++ libraries/material-networking/src/material-networking/KTXCache.cpp
