+++ libraries/material-networking/src/material-networking/KTXCache.h
