+++ libraries/material-networking/src/material-networking/MaterialNetworkingLogging.h
@@ -1,0 +1,11 @@
+//
+//  Created by Sam Gondelman on 2/7/2019
+//  Copyright 2019 High Fidelity, Inc.
+//
+//  Distributed under the Apache License, Version 2.0.
+//  See the accompanying file LICENSE or http://www.apache.org/licenses/LICENSE-2.0.html
+//
+
+#include <QLoggingCategory>
+
+Q_DECLARE_LOGGING_CATEGORY(materialnetworking)
