+++ libraries/material-networking/src/material-networking/ShaderCache.cpp
