+++ libraries/material-networking/src/material-networking/TextureCache.h
