+++ libraries/material-networking/src/material-networking/TextureCacheScriptingInterface.cpp
