+++ libraries/material-networking/src/material-networking/TextureCacheScriptingInterface.h
