+++ libraries/model-baker/CMakeLists.txt
@@ -1,4 +1,5 @@
 set(TARGET_NAME model-baker)
 setup_hifi_library()
 
-link_hifi_libraries(shared task gpu graphics hfm)
+link_hifi_libraries(shared shaders task gpu graphics hfm material-networking)
+include_hifi_library_headers(networking)
