+++ libraries/physics/CMakeLists.txt
@@ -7,6 +7,7 @@ include_hifi_library_headers(avatars)
 include_hifi_library_headers(audio)
 include_hifi_library_headers(octree)
 include_hifi_library_headers(animation)
+include_hifi_library_headers(material-networking)
 include_hifi_library_headers(model-networking)
 include_hifi_library_headers(image)
 include_hifi_library_headers(ktx)
