+++ libraries/procedural/CMakeLists.txt
@@ -1,4 +1,4 @@
 set(TARGET_NAME procedural)
 setup_hifi_library()
-link_hifi_libraries(shared gpu shaders networking graphics model-networking ktx image)
+link_hifi_libraries(shared gpu shaders networking graphics material-networking ktx image)
